Add Stats struct and PrintStats() function
This commit adds a Stats struct to store statistical data and a
PrintStats() function to print the data. The Stats struct includes
members to hold the total number of files/directories, the number of
missing files/directories, the number of fixed/unfixed files/directories.
The PrintStats() function prints the statistical data.