Add statistics tracking and reporting
This commit adds statistics tracking and reporting to the program.
The `FixFileTime()` function now tracks the number of fixed/unfixed
files/directories in the Stats struct. The `ProcessDirsRecursively()`
function now tracks the total number of files/directories in the source
directory as well as the number of missing files/directories in the
destination directory. The `PrintStats()` function prints the
statistical data.