Fix logic error in UpdateFileTimeIfEarlier()
This commit fixes a logic error in the UpdateFileTimeIfEarlier()
function that caused an error when opening destination files/directories
with different access rights. The function now correctly determines
whether the destination file is a file or a directory before calling
FixFileTime(). The commit also adds comments to the
UpdateFileTimeIfEarlier() function.