Change order of parsing command line arguments
The new implementation of wmain() function changes the order of setting
program options and parsing command line arguments. If there are more
than three command line arguments, it sets program options from command
line arguments first and then assigns values to sourcePath and destPath.
Otherwise, it assigns values to sourcePath and destPath directly.
This change does not affect the functionality of the program as long as
the correct number of command line arguments are provided.