Add FileOrDirExists() function
This commit adds a new function called FileOrDirExists() in utils.cpp
that checks if a file or directory exists. In wmain() function,
it checks if the source and destination file or directory exist
by calling the FileOrDirExists() function. Previously, the existence
of files and directories was checked in ProcessFilesOrDirs() function
anyway. Now it is checked before this function is called.